In short, nimhax with ctr-httpwn
Move unused ctr-httpwn files to a separate folder.
Update ctr-httpwn code to use specific result codes.
Wrote fileio workaround for sdmc on xmltakeover runs,
because for some reason (ctrulib 2.0?) fopen fails.
Wrote rops for nim:s 0x5E as httpwn replaces
user-agent on a predetermined stack buffer offset
and pivot to rops set over by IPC_Desc_StaticBuffer.
Create an linear flow fake file service for am:net.
Wrote am rop for fake file service to pass to AM.
zoogie updated httpwn rop gen code for httpwn >=11.4
using TuxSH's universal-otherapp code as reference
of httpwn >=11.4 code.

+++ source/ctr_httpwn/config.cpp
@@ -152,7 +152,7 @@ int config_parse(configctx *config, char *xml)
 					ret = AM_GetTitleInfo((FS_MediaType)tmpval0, 1, &titleid, &title_entry);
 					if(ret!=0)
 					{
-						printf("WARNING: Failed to get the title info for the title specified by config: 0x%08x.\n", (unsigned int)ret);
+						printf("WARNING: Failed to get the title info for the title specified by config: 0x%08x.\n", ret);
 						if(textptr)printf("targeturl name = %s\n", textptr);
 						printf("Config won't be loaded for this targeturl entry.\n");
 						continue;
